pyproject.toml
[tool.poetry]
name = "netflix-search+ddd-backend"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
requests = "^2.28.1"
types-requests = "^2.28.11.5"
python-dotenv = "^0.21.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
pyjwt = "^2.6.0"
django-url-filter = "^0.3.15"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}



[tool.poetry.dev-dependencies]
mypy = "^0.991"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
